% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym_mbpp2.R
\name{asym_mbpp2}
\alias{asym_mbpp2}
\title{Fit asymptotic approximation of hierarchical N-mixture
model for estimation of northern fur seal pup production.}
\usage{
asym_mbpp2(det_formula = ~rcode * resample * observer,
  avail_formula = ~rcode * resample, mark_data, resight_data, par, ...)
}
\arguments{
\item{det_formula}{formula for the detection model}

\item{avail_formula}{formula for the avaialbility model}

\item{mark_data}{Data frame providing pups marked and dead pup counts
for each site}

\item{resight_data}{Data frame providing marked and unmarked resight counts}

\item{par}{Optional start value specification.}

\item{...}{additional arguments passed to nlminb() or TMB::MakeADFun()
for optimization}
}
\description{
Fit asymptotic approximation of hierarchical N-mixture
model for estimation of northern fur seal pup production.
}
\author{
Devin S. Johnson
}
