% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_asym.R
\name{boot_asym}
\alias{boot_asym}
\title{Parameteric bootstrap sample for abundance}
\usage{
boot_asym(object, size = 10000)
}
\arguments{
\item{object}{A fitted model object from aysm_mbpp}

\item{size}{Number of bootstrap samples to draw}
}
\description{
Parameteric bootstrap sample for abundance
}
\details{
This function uses the asymtotic model fit from TMB to generate
a bootstrap sample for parameter inference
}
