% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{compile_mbpp_tmb}
\alias{compile_mbpp_tmb}
\title{Compile package TMB function}
\usage{
compile_mbpp_tmb()
}
\arguments{
\item{model}{Details which model to compile}
}
\description{
Compile package TMB function
}
\details{
This function is only used once after the
package is installed to compile the c++ TMB code for
asymptotic abundance estimation. Simply call the function
to compile the source code.
}
