% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_mbpp.R
\name{jags_mbpp}
\alias{jags_mbpp}
\title{Fit asymptotic approximation of hierarchical N-mixture
model for estimation of northern fur seal pup production.}
\usage{
jags_mbpp(det_formula = ~rcode * resample * observer,
  avail_formula = ~rcode, mark_data, resight_data, par, ...)
}
\arguments{
\item{det_formula}{formula for the detection model}

\item{avail_formula}{formula for the avaialbility model}

\item{mark_data}{Data frame providing pups marked and dead pup counts
for each site}

\item{resight_data}{Data frame providing marked and unmarked resight counts}

\item{par}{Optional start value specification.}

\item{...}{additional arguments passed to \link{R2jags::jags} such as
n.chains, n.iter, n.burnin, or n.thin
for optimization}
}
\description{
Fit asymptotic approximation of hierarchical N-mixture
model for estimation of northern fur seal pup production.
}
\author{
Devin S. Johnson
}
