% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{nfs_mark}
\alias{nfs_mark}
\title{Northern fur seal shear marking data from 2016}
\format{A data frame with 19 observations on the following 6 variables.

\describe{
\item{icode}{Island designation}

\item{rcode}{Rookery designation}

\item{deadpups}{Number of dead pups counted on survey}

\item{pupsborn}{Estimated pup production using design-based method}

\item{sep}{Estimated standard error using design-based method}

\item{M}{Number of pups shear marked}
}}
\source{
Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA 7600 Sand
Point Way NE Seattle, WA 98115
}
\description{
Northern fur seal shear marking data from 2016
}
\keyword{datasets}
