% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{nfs_resight}
\alias{nfs_resight}
\title{Northern fur seal mark resight data from 2016}
\format{
A data frame with 76 observations on the following 6 variables.

\describe{
\item{icode}{Island designation}

\item{rcode}{Rookery designation}

\item{resample}{The resight occasion designation}

\item{observer}{The observer designation}

\item{m}{Number of marked pups observed}

\item{u}{Number of unmarked pups observed}
}
}
\source{
Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA 7600 Sand
Point Way NE Seattle, WA 98115
}
\description{
Northern fur seal mark resight data from 2016
}
\keyword{datasets}
